move inside components to separate files
