+++ src/components/BodyParts/BodyParts.js
@@ -1,4 +1,5 @@
 import "./BodyParts.css";
+import Areas from "../Areas/Areas";
 
 function BodyParts({ handleChange, points }) {
   const frontParts = [
@@ -20,23 +21,6 @@ function BodyParts({ handleChange, points }) {
     ["right-leg-back", 9],
   ];
 
-  function Areas({ parts, side }) {
-    const areas = parts.map((part, index) => {
-      index = side === "back" ? index + 7 : index;
-      let isChecked = points[index] === 0 ? "false" : "true";
-      return (
-        <div
-          className={`body-part ${part[0]} ${
-            isChecked === "true" ? "checked" : ""
-          }`}
-          data-value={isChecked === "false" ? part[1] : 0}
-          onClick={(e) => handleChange(e, index)}
-        ></div>
-      );
-    });
-    return areas;
-  }
-
   return (
     <section>
       <h2>Część A: Powierzchnia skóry zajęta przez AZS</h2>
@@ -49,11 +33,21 @@ function BodyParts({ handleChange, points }) {
       </section>
       <div className="front-body">
         <h3>Przód ciała</h3>
-        <Areas parts={frontParts} points={points} side={"front"} />
+        <Areas
+          parts={frontParts}
+          points={points}
+          side={"front"}
+          handleChange={handleChange}
+        />
       </div>
       <div className="back-body">
         <h3>Tył ciała</h3>
-        <Areas parts={backParts} points={points} side={"back"} />
+        <Areas
+          parts={backParts}
+          points={points}
+          side={"back"}
+          handleChange={handleChange}
+        />
       </div>
     </section>
   );
